Removing test functions for official release
