+++ setup.cfg
@@ -28,7 +28,6 @@ console_scripts =
     drupal_backup_database.py = tubular.scripts.drupal_backup_database:backup_database
     drupal_clear_varnish.py = tubular.scripts.drupal_clear_varnish:clear_varnish_cache
     drupal_deploy.py = tubular.scripts.drupal_deploy:deploy
-    drupal_deploy_test.py = tubular.scripts.drupal_deploy_test:deploy
     drupal_fetch_deployed_tag.py = tubular.scripts.drupal_fetch_deployed_tag:fetch_deployed_tag
     find_and_advance_pipeline.py = tubular.scripts.find_and_advance_pipeline:find_and_advance_pipeline
     jenkins_trigger_build.py = tubular.scripts.jenkins_trigger_build:trigger
