+++ tubular/scripts/drupal_deploy.py
@@ -16,13 +16,13 @@ sys.path.append(path.dirname(path.dirname(path.abspath(__file__))))
 from tubular import drupal  # pylint: disable=wrong-import-position
 
 
-# TODO: LEARNER-3645 Copy the contents of drupal_deploy_test.py to here and REMOVE drupal_deploy_test.py
 @click.command()
 @click.option("--env", help="The environment to deploy code in.", type=str, required=True)
 @click.option("--username", help="The Acquia username necessary to run the command.", type=str, required=True)
 @click.option("--password", help="The Acquia password necessary to run the command.", type=str, required=True)
-@click.option("--tag", help="The tag name to be deployed to the environment.", type=str, required=True)
-def deploy(env, username, password, tag):
+@click.option("--branch_or_tag", help="The branch or tag name to be deployed to the environment.",
+              type=str, required=True)
+def deploy(env, username, password, branch_or_tag):
     """
     Deploys a given tag to the specified environment.
 
@@ -30,9 +30,9 @@ def deploy(env, username, password, tag):
         env (str): The environment to deploy code in (e.g. test or prod)
         username (str): The Acquia username necessary to run the command.
         password (str): The Acquia password necessary to run the command.
-        tag (str): The tag to deploy to the specified environment.
+        branch_or_tag (str): The branch or tag to deploy to the specified environment.
     """
-    drupal.deploy(env, username, password, tag)
+    drupal.deploy(env, username, password, branch_or_tag)
 
 if __name__ == "__main__":
     deploy()  # pylint: disable=no-value-for-parameter
